Move local-var-related types into their own header.
